requirements.txt
Python 			3.8.2
asgiref                 3.2.3
astroid                 2.3.3
colorama                0.4.3
Django                  3.0.3
django-crispy-forms     1.9.0
django-multiselectfield 0.1.12
flake8                  3.8.3
isort                   4.3.21
lazy-object-proxy       1.4.3
mccabe                  0.6.1
Pillow                  7.0.0
pip                     20.0.2
pycodestyle             2.6.0
pyflakes                2.2.0
pylint                  2.4.4
pytz                    2019.3
setuptools              45.2.0
six                     1.14.0
sqlparse                0.3.1
wheel                   0.34.2
wrapt                   1.11.2